refactor: reorganize message handler types and improve error formatting interfaces
